+++ radios/code/favealbums.js
@@ -17,12 +17,12 @@ var faveAlbums = function() {
                     addTracks();
                 } else {
                     infobar.notify(infobar.NOTIFY,language.gettext('label_gotnotracks'));
-                    playlist.radioManager.stop();
+                    playlist.radioManager.stop(null);
                 }
             },
             fail: function() {
                 infobar.notify(infobar.NOTIFY,language.gettext('label_gotnotracks'));
-                playlist.radioManager.stop();
+                playlist.radioManager.stop(null);
                 populated = false;
             }
         });
@@ -68,14 +68,14 @@ var faveAlbums = function() {
             } else {
                 debug.warn("SMARTPLAYLIST","Got NO tracks",data);
                 infobar.notify(infobar.NOTIFY,language.gettext('label_gotnotracks'));
-                playlist.radioManager.stop();
+                playlist.radioManager.stop(null);
                 running = false;
             }
         },
 
         Fail: function() {
             infobar.notify(infobar.NOTIFY,language.gettext('label_gotnotracks'));
-            playlist.radioManager.stop();
+            playlist.radioManager.stop(null);
             populated = false;
             running = false;
         }