+++ radios/code/lastfmtrackradio.js
@@ -160,7 +160,7 @@ var lastFMTrackRadio = function() {
 			} else {
 				if (currpage == 1) {
 					infobar.notify(infobar.ERROR, "Got no data from Last.FM");
-					playlist.radioManager.stop();
+					playlist.radioManager.stop(null);
 				}
 			}
 		},