+++ radios/code/starRadios.js
@@ -58,14 +58,14 @@ var starRadios = function() {
             } else {
                 debug.warn("SMARTPLAYLIST","Got NO tracks",data);
                 infobar.notify(infobar.NOTIFY,language.gettext('label_gotnotracks'));
-                playlist.radioManager.stop();
+                playlist.radioManager.stop(null);
                 running = false;
             }
         },
 
         Fail: function() {
             infobar.notify(infobar.NOTIFY,language.gettext('label_gotnotracks'));
-            playlist.radioManager.stop();
+            playlist.radioManager.stop(null);
             populating = false;
             running = false;
         },